decline_join_request/requirements.txt
google-cloud-logging
google-python-cloud-debugger
functions-framework==3.*
python-dotenv
requests
